requirements_mac.txt
comtypes==1.4.4
matplotlib==3.8.4
numpy==1.26.4
PyAutoGUI==0.9.54
pycaw==20240210
pynput==1.7.7
PyQt5==5.15.10
PyQt5_sip==12.13.0
scipy==1.13.1
pyobjc-core
pyobjc-framework-CoreAudio
